requirements.txt
fastapi
uvicorn
sentence-transformers
numpy
faiss-cpu
python-docx